Initial version of a redesigned FX Mixer
The meters are now rendered in code (using a rectangle with a gradient
brush), i.e. they can now be rendered at any size and do not rely on
static pixmaps anymore.

The channel name is now displayed at the bottom and the text is elided
if it is too long. The tooltip of the text shows the full name.

Replaced the pixmaps of the solo and mute button and got rid of the
LCD to display the channel numbers.

The send code was in part commented out to make the code work. For a
complete version it will have to be reinstated somehow. Due to these
changes this version currently crashes on exit.

TODOs: Consider showing the amplitude in DB in the meters. Some more
layouting.