Changes the way the widgets are layouted in the main toolbar
Adds a new horizontal layout to MainWindow which acts as the main
layout. The old grid layout is added to this layout as the first layout.
Only the toolbuttons are added to the grid layout. Everything else is
added to the main layout.

The methods MainWindow::addWidgetToToolBar and
MainWindow::addSpacingToToolBar have been removed. Instead it is now
possible to fetch the horizontal layout by calling the new method
MainWindow::getToolbarMainLayout. This is currently only done in
SongEditor.cpp but it seems that this is the wrong place anyway and it
should be changed later.

Removed the setting of the maximum height from TimeDisplayWidget.
Switched the setting of the font size for the time display to using
pixels as the toolbar is currently set in pixels as well.

Renamed several local variables to make the code better understandable.