Bug fix: Race condition causing crash
	Thanks to Dom for helping me figure this one out!
	When clicking the clip indicators repeatedely to reset them
while the track was playing and clipping, we could eventually crash
LMMS. The reason behind it was a race condition, caused by using
AutomatableModel::setValue instead of
AutomatableModel::setAutomatableValue. The former also adds a checkpoint
to the Journalling, which is meant to be used by the UI (which is single
threaded), hence it's not multi-thread safe. Processing threads should
use setAutomatableValue instead which doesn't journal the changes and
should be safe to use with multithreads. Places that used setValue were
changed to use setAutomatableValue insted.