First version of LmmsComboBox
Add a first version of the class `LmmsComboBox` which is a combo box that inherits from `QComboBox` and that knows how to deal with `ComboBoxModel`.

The intention is that this class supersedes the class `ComboBox` in all contexts where no automation is needed/wanted. It should be styled using Qt style sheets later.

Another advantage of inheriting from `QComboBox` is that it should automatically know how to "behave" with regards to HiDPI displays.